Match the structure of the module tree with the API path.
This commit determines how to handle the `:id` parameter of the paths.
